so this are my codes from my practice session in coursera.com
in Lecture 10 which happens to be an "img" display tag script my image wasn't able to upload for reasons not yet  clear to me. maybe it's because i don't have the pictures on my desktop but expected my web browser to fetch it for  me? maybe if you're reading this kindly check on it and make a reply.